cos/admission.rs: colocate 20 ECN-admission + flow-share tests with production (#984 P3 Phase 5b)
Append 20 tests to the existing cos/admission.rs::tests block:
- 15 admission_*/admission_ecn_* tests covering apply_cos_admission_ecn_policy
- 5 flow_share_limit_* tests covering cos_queue_flow_share_limit

Imports added: PROTO_TCP, PreparedTxRecycle, PreparedTxRequest,
ECN_CE, ECN_ECT_0, ECN_MASK, ECN_NOT_ECT.

Also drops 11 stale cfg-test imports from tx/mod.rs (Codex round-1
cleanup): bdp_floor_bytes, COS_ECN_MARK_THRESHOLD_DEN/_NUM,
COS_FLOW_FAIR_MIN_SHARE_BYTES, ECN_CE/ECT_0/MASK/NOT_ECT,
apply_cos_admission_ecn_policy, cos_flow_aware_buffer_limit,
cos_queue_flow_share_limit. Kept cos_item_flow_key in production
import — tx/cos_classify.rs reaches it via use super::*.

Rebased onto current master after #1009 merged.
cargo test --bins 865/0/2.
